requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor==3.6.0
pymongo[srv]==4.9.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.0
python-multipart==0.0.12
